# public/openapi/v1.yaml
openapi: 3.0.3
info:
  title: SICA B2B Catalog API
  version: '1.0.0'
  description: |
    Public catalog API for external integrators (affiliate sites, sister
    platforms, recruitment partners). Read-only access to SICA's
    university and program data, plus outbound webhooks for change
    notifications.

    ## Authentication
    Every request needs an `Authorization: Bearer sk_live_…` header.
    Issue a key from `/admin/api-keys` (admin-only). The plaintext is
    shown exactly once on creation; we only keep the SHA-256 hash.

    ## Rate limits
    Per-key token bucket, capacity = the limit set on the key
    (default 100 req/min, configurable per key at issue time). On
    overflow: HTTP 429 with `Retry-After`. Every response includes
    `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`.

    ## Caching
    Responses are cached in-memory for 10 min, per (api_key_id,
    query_string). Cache status: `X-Cache: HIT | MISS | BYPASS`. Pass
    `?nocache=1` to bypass.

    ## Webhooks
    Subscribe to `university.*` and `program.*` events with
    `/v1/webhooks`. Every delivery is HMAC-SHA256 signed with your
    per-subscription secret (`X-SICA-Signature`). Verify with
    `timingSafeEqual` / `hmac.compare_digest`. At-least-once delivery
    — dedupe on `X-SICA-Delivery-Id`.

    ## CORS (browser callers)
    Browser-side consumers (a partner's public site calling SICA via
    `fetch()` from JS) need CORS headers. Set the allowlist per key
...

...
    ships.

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /v1/webhooks:
    get:
      tags: [webhooks]
      summary: List webhook subscriptions
      description: |
        Returns all subscriptions owned by the authenticated API key,
        newest first. The `secret` is never returned here — it is
        only shown once at create time.
      operationId: listWebhookSubscriptions
      responses:
        '200':
          description: Subscriptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items: { $ref: '#/components/schemas/WebhookSubscription' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }

    post:
      tags: [webhooks]
      summary: Create a webhook subscription
      description: |
        Register a URL to receive event POSTs. The `secret` is
        generated server-side and returned in the response — this is
        the only time the secret will be shown. Store it server-side
        and use it to verify the `X-SICA-Signature` header on every
        incoming delivery.
      operationId: createWebhookSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                  description: 'https:// URL (http://localhost allowed for dev)'
                events:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items: { $ref: '#/components/schemas/WebhookEvent' }
                description:
                  type: string
                  maxLength: 200
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                type: object
                required: [subscription, secret, secret_note]
                properties:
                  subscription: { $ref: '#/components/schemas/WebhookSubscription' }
                  secret:
                    type: string
                    description: 'HMAC-SHA256 secret. Shown exactly once.'
                  secret_note:
                    type: string
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/webhooks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    patch:
      tags: [webhooks]
      summary: Update a subscription
      description: |
        Update the URL, subscribed events, description, or active state.
        The `secret` is NOT updatable — to rotate the signing secret,
        create a new subscription and delete the old one. Updates are
        scoped to the authenticated key (a key cannot edit another
        key's subscriptions).
      operationId: updateWebhookSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items: { $ref: '#/components/schemas/WebhookEvent' }
                description:
                  type: string
                  maxLength: 200
                  nullable: true
                active: { type: boolean }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription: { $ref: '#/components/schemas/WebhookSubscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }

    delete:
      tags: [webhooks]
      summary: Disable a subscription (soft-revoke)
      description: |
        Sets `active = false` on the subscription. The row stays for
        audit; deliveries stop immediately. Re-enable by PATCHing
        `active: true`. Scoped to the authenticated key.
      operationId: deleteWebhookSubscription
      responses:
        '200':
          description: Disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      active: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sk_live_<base64>
      description: |
        API key in `Authorization: Bearer sk_live_…` format. Issue a
        key from `/admin/api-keys`. Required scope: `read:catalog`.
...
  headers:
...
    XRateLimitLimit:
      description: Per-key request limit (per minute).
      schema:
        type: integer
    XRateLimitRemaining:
      description: Tokens remaining in the current bucket.
      schema:
        type: integer
    XRateLimitReset:
      in: header
      name: X-RateLimit-Reset
      description: 'Seconds until the token bucket refills enough to allow another request.'
      schema: { type: integer }
...
  responses:
    BadRequest:
      description: Invalid query parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, malformed, revoked, or expired API key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Rate limit exceeded. Retry after the `Retry-After` seconds.
      headers:
        Retry-After:
          description: Seconds until the next token.
          schema:
            type: integer
        X-RateLimit-Limit: { $ref: '#/components/headers/XRateLimitLimit' }
        X-RateLimit-Remaining: { $ref: '#/components/headers/XRateLimitRemaining' }
        X-RateLimit-Reset: { $ref: '#/components/headers/XRateLimitReset' }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    WebhookEvent:
      type: string
      enum:
        - university.created
        - university.updated
        - university.deleted
        - program.created
        - program.updated
        - program.deleted
      description: |
        Event identifier. Emitted as the value of the `X-SICA-Event`
        header on every delivery and inside the JSON payload's
        `event` field.
    WebhookSubscription:
      type: object
      required: [id, url, events, active, created_at, success_count, failure_count]
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        events:
          type: array
          items: { $ref: '#/components/schemas/WebhookEvent' }
        description:
          type: string
          nullable: true
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        last_triggered_at:
          type: string
          format: date-time
          nullable: true
        success_count: { type: integer, minimum: 0 }
        failure_count: { type: integer, minimum: 0 }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string, example: 'Missing or malformed Authorization header.' }
...